requirements.txt
openai
fastapi
uvicorn
pydantic
python-dotenv
tavily-python